* added some join/host/change mod code * added a "default" mod to the mod list in the mods window (in order to be able to unload all mods and get back to default) * minor fixes

+++ src/ta3d/src/gfx/gui/object.h
@@ -59,7 +59,7 @@ namespace Gui
 		/*!
 		** \brief
 		*/
-		uint32  num_entries() const {return Text.size();}
+		uint32  num_entries() const {return uint32(Text.size());}
 
 		/*!
 		** \brief Reacts to a message transfered from the interface
